service端数据类型转义修正，intval修正为floatval 0626 1751

+++ app/api/YjfpService.class.php
@@ -254,22 +254,22 @@ class api_YjfpService extends api_Abstract implements YjfpServiceIf
             //  ['bd',$com_title['bd']]    ['pm',$com_title['pm']]     ['delivery',$com_title['delivery']]
             //线索提供人 = 有效线索提供
             $re[0] = $uid_arr_final[$invoice['line_role']];
-            $re[0]['bli'] = intval($sys['effective_clue']);
+            $re[0]['bli'] = floatval($sys['effective_clue']);
             $re[0]['money'] = $re[0]['bli'] * floatval($invoice['money']) * 0.01;
             $re[0]['title'] = ['clue',$com_title['clue']];
             //BD = 合同签订 + 回款
             $re[1] = $uid_arr_final[$invoice['bd_role']];
-            $re[1]['bli'] = intval($sys['contract_sign']) + intval($sys['receivable']);
+            $re[1]['bli'] = floatval($sys['contract_sign']) + intval($sys['receivable']);
             $re[1]['money'] = $re[1]['bli'] * floatval($invoice['money']) * 0.01;
             $re[1]['title'] = ['bd',$com_title['bd']];
             //项目经理 = 立项需求表、企业项目对接
             $re[2] = $uid_arr_final[$invoice['project_role']];
-            $re[2]['bli'] = intval($sys['project_docking']);
+            $re[2]['bli'] = floatval($sys['project_docking']);
             $re[2]['money'] = $re[2]['bli'] * floatval($invoice['money']) * 0.01;
             $re[2]['title'] = ['pm',$com_title['pm']];
             //交付人 = 候选人简历 + 候选人意向沟通、简历报告制作 + 候选人推荐及面试更进 + 候选人薪酬及offer谈判 + 候选人背景调查、入职更进
             $re[3] = $uid_arr_final[$merge_clear[0]];
-            $re[3]['bli'] = intval($sys['resume_provision']) + intval($sys['intention_communicate']) + intval($sys['interview_follow']) + intval($sys['offer_negotiate']) + intval($sys['reference_check']);
+            $re[3]['bli'] = floatval($sys['resume_provision']) + floatval($sys['intention_communicate']) + floatval($sys['interview_follow']) + floatval($sys['offer_negotiate']) + floatval($sys['reference_check']);
             $re[3]['money'] = $re[3]['bli'] * floatval($invoice['money']) * 0.01;
             $re[3]['title'] = ['delivery',$com_title['delivery']];
             if(count($merge_clear)>1){
@@ -288,12 +288,12 @@ class api_YjfpService extends api_Abstract implements YjfpServiceIf
             $re2[0]['title'] = ['clue',$com_title2['clue']];
             //2、BD--合同签订
             $re2[1] = $uid_arr_final[$invoice['bd_role']];
-            $re2[1]['bli'] = intval($sys['contract_sign']);
+            $re2[1]['bli'] = floatval($sys['contract_sign']);
             $re2[1]['money'] = $re2[1]['bli'] * floatval($invoice['money']) * 0.01;
             $re2[1]['title'] = ['contract',$com_title2['contract']];
             //3、回款
             $re2[2] = $uid_arr_final[$invoice['bd_role']];
-            $re2[2]['bli'] = intval($sys['receivable']);
+            $re2[2]['bli'] = floatval($sys['receivable']);
             $re2[2]['money'] = $re2[2]['bli'] * floatval($invoice['money']) * 0.01;
             $re2[2]['title'] = ['receivable',$com_title2['receivable']];
             //4、立项需求表企业项目对接
@@ -303,14 +303,14 @@ class api_YjfpService extends api_Abstract implements YjfpServiceIf
             if(count($jf_cc['items'])>0){
                 $re2[4] = $uid_arr_final[$jf_cc['items'][0]['role_id']];
             }
-            $re2[4]['bli'] = intval($sys['resume_provision']);
+            $re2[4]['bli'] = floatval($sys['resume_provision']);
             $re2[4]['money'] = $re2[4]['bli'] * floatval($invoice['money']) * 0.01;
             $re2[4]['title'] = ['resume',$com_title2['resume']];
             //6、顾问面试  候选人意向沟通、简历报告制作        取CC第一个
             if(count($jf_cc['items'])>0){
                 $re2[5] = $uid_arr_final[$jf_cc['items'][0]['role_id']];
             }
-            $re2[5]['bli'] = intval($sys['intention_communicate']);
+            $re2[5]['bli'] = floatval($sys['intention_communicate']);
             $re2[5]['money'] = $re2[5]['bli'] * floatval($invoice['money']) * 0.01;
             $re2[5]['title'] = ['intention',$com_title2['intention']];
             /*******************************************************************/
@@ -358,17 +358,17 @@ class api_YjfpService extends api_Abstract implements YjfpServiceIf
                 }
 
             }
-            $re2[6]['bli'] = intval($sys['interview_follow']);
+            $re2[6]['bli'] = floatval($sys['interview_follow']);
             $re2[6]['money'] = $re2[6]['bli'] * floatval($invoice['money']) * 0.01;
             $re2[6]['title'] = ['recommend',$com_title2['recommend']];
             //8、薪酬offer谈判
             $re2[7] = $uid_arr_final[$jf_offer['items'][0]['role_id']];
-            $re2[7]['bli'] = intval($sys['offer_negotiate']);
+            $re2[7]['bli'] = floatval($sys['offer_negotiate']);
             $re2[7]['money'] = $re2[7]['bli'] * floatval($invoice['money']) * 0.01;
             $re2[7]['title'] = ['offer',$com_title2['offer']];
             //7、候选人背景调查 入职跟进
             $re2[8] = $uid_arr_final[$jf_rz['items'][0]['role_id']];
-            $re2[8]['bli'] = intval($sys['reference_check']);
+            $re2[8]['bli'] = floatval($sys['reference_check']);
             $re2[8]['money'] = $re2[8]['bli'] * floatval($invoice['money']) * 0.01;
             $re2[8]['title'] = ['entry',$com_title2['entry']];
             $invoice['info']['process'] = $re2;
